+++ src/main/java/com/konkuk/Eodikase/domain/data/service/CourseDataService.java
@@ -1,10 +1,12 @@
 package com.konkuk.Eodikase.domain.data.service;
 
-import com.konkuk.Eodikase.domain.data.dto.request.FilteredCourseDataCountRequest;
-import com.konkuk.Eodikase.domain.data.dto.request.FilteredCourseDataRequest;
-import com.konkuk.Eodikase.domain.data.dto.response.*;
 import com.konkuk.Eodikase.domain.data.entity.*;
 import com.konkuk.Eodikase.domain.data.repository.*;
+import com.konkuk.Eodikase.dto.request.FilteredCourseDataCountRequest;
+import com.konkuk.Eodikase.dto.request.FilteredCourseDataRequest;
+import com.konkuk.Eodikase.dto.response.*;
+import com.konkuk.Eodikase.dto.response.data.entity.*;
+import com.konkuk.Eodikase.dto.response.data.repository.*;
 import com.konkuk.Eodikase.domain.member.repository.MemberRepository;
 import com.konkuk.Eodikase.exception.badrequest.InvalidDataOrderException;
 import com.konkuk.Eodikase.exception.badrequest.InvalidRegionException;
