+++ src/main/java/com/konkuk/Eodikase/domain/member/controller/MemberController.java
@@ -3,6 +3,10 @@ package com.konkuk.Eodikase.domain.member.controller;
 import com.konkuk.Eodikase.domain.member.dto.request.*;
 import com.konkuk.Eodikase.domain.member.dto.response.*;
 import com.konkuk.Eodikase.domain.member.service.MemberService;
+import com.konkuk.Eodikase.dto.request.*;
+import com.konkuk.Eodikase.dto.request.response.*;
+import com.konkuk.Eodikase.dto.response.*;
+import com.konkuk.Eodikase.dto.response.response.*;
 import com.konkuk.Eodikase.security.auth.LoginUserId;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
