+++ src/main/java/com/konkuk/Eodikase/dto/request/MemberProfileUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.domain.member.dto.request;
+package com.konkuk.Eodikase.dto.request;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
