+++ src/main/java/com/konkuk/Eodikase/dto/response/OAuthMemberSignUpResponse.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.domain.member.dto.response;
+package com.konkuk.Eodikase.dto.response;
 
 import lombok.*;
