+++ src/main/java/com/konkuk/Eodikase/dto/response/OAuthTokenResponse.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.domain.auth.dto.response;
+package com.konkuk.Eodikase.dto.response;
 
 import lombok.*;
