+++ src/test/java/com/konkuk/Eodikase/service/MemberServiceTest.java
@@ -1,10 +1,10 @@
 package com.konkuk.Eodikase.service;
 
-import com.konkuk.Eodikase.domain.member.dto.request.MemberProfileUpdateRequest;
-import com.konkuk.Eodikase.domain.member.dto.request.MemberSignUpRequest;
-import com.konkuk.Eodikase.domain.member.dto.request.ResetPasswordRequest;
-import com.konkuk.Eodikase.domain.member.dto.request.OAuthMemberSignUpRequest;
-import com.konkuk.Eodikase.domain.member.dto.request.PasswordVerifyRequest;
+import com.konkuk.Eodikase.dto.request.MemberProfileUpdateRequest;
+import com.konkuk.Eodikase.dto.request.MemberSignUpRequest;
+import com.konkuk.Eodikase.dto.request.ResetPasswordRequest;
+import com.konkuk.Eodikase.dto.request.OAuthMemberSignUpRequest;
+import com.konkuk.Eodikase.dto.request.PasswordVerifyRequest;
 import com.konkuk.Eodikase.domain.member.dto.response.*;
 import com.konkuk.Eodikase.domain.member.entity.Member;
 import com.konkuk.Eodikase.domain.member.entity.MemberPlatform;
@@ -13,6 +13,9 @@ import com.konkuk.Eodikase.domain.member.entity.MemberProfileImage;
 import com.konkuk.Eodikase.domain.member.repository.MemberProfileImageRepository;
 import com.konkuk.Eodikase.domain.member.repository.MemberRepository;
 import com.konkuk.Eodikase.domain.member.service.MemberService;
+import com.konkuk.Eodikase.dto.request.response.*;
+import com.konkuk.Eodikase.dto.response.*;
+import com.konkuk.Eodikase.dto.response.response.*;
 import com.konkuk.Eodikase.exception.badrequest.*;
 import com.konkuk.Eodikase.exception.notfound.NotFoundMemberException;
 import com.konkuk.Eodikase.support.AwsS3Uploader;
